XRDDEV-86 Move version number to source code
