+++ src/build.gradle
@@ -30,7 +30,7 @@ sonarqube {
         property "sonar.projectKey", "xroad"
         property "sonar.projectName", "X-Road"
         property "sonar.projectDescription", "Data Exchange Layer"
-        property "sonar.projectVersion", XROAD_VERSION
+        property "sonar.projectVersion", "6.19.0"
     }
 }
