+++ src/gradle.properties
@@ -2,4 +2,3 @@
 // We need to enable access, so that we can refer to external schemas when
 // generating classes with JAXB.
 systemProp.javax.xml.accessExternalSchema=all
-XROAD_VERSION=6.19.0
