+++ src/proxy-ui/app/controllers/about_controller.rb
@@ -21,10 +21,10 @@
 # THE SOFTWARE.
 #
 
-java_import Java::ee.ria.xroad.common.util.JarUtils
+java_import Java::ee.ria.xroad.common.Version
 
 class AboutController < ApplicationController
     def index
-        @version = JarUtils.readJarManifestProperty(JarUtils::COMMON_UTIL_JAR_PATH, JarUtils::IMPLEMENTATION_VERSION)
+        @version = Version::XROAD_VERSION
     end
 end
